requirements.txt
numpy>=1.19.0,<1.24.0
scipy>=1.7.0,<1.10.0
matplotlib>=3.4.0,<3.7.0
torch>=1.9.0,<2.0.0
torchvision>=0.10.0,<0.15.0
opencv-python>=4.5.0,<4.8.0
pillow>=8.0.0,<10.0.0
tqdm>=4.62.0,<4.65.0
pyyaml>=5.4.0,<6.0.0
jupyter>=1.0.0,<1.1.0
notebook>=6.4.0,<7.0.0
pytest>=6.2.0,<7.0.0
black>=21.7b0,<22.0.0
flake8>=3.9.0,<4.0.0
pre-commit>=2.15.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
requests>=2.26.0,<3.0.0
tensorboard>=2.6.0,<3.0.0
wandb>=0.12.0,<0.13.0